requirements.txt
joblib
seaborn
pandas
matplotlib.pyplot
streamlit
numpy